backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "arbitragevault-backend"
version = "0.1.0"
description = "ArbitrageVault backend API for book arbitrage analysis"
authors = [{name = "Aziz Traore", email = "aziztraorebf@gmail.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0,<0.105.0",
    "uvicorn[standard]>=0.24.0,<0.25.0", 
    "pydantic-settings>=2.1.0,<2.2.0",
    "sqlalchemy[asyncio]>=2.0.23,<2.1.0",
    "psycopg[binary]>=3.1.0,<3.2.0",
    "alembic>=1.13.0,<1.14.0",
    "python-jose[cryptography]>=3.3.0,<3.4.0",
    "passlib[argon2,bcrypt]>=1.7.4,<1.8.0",
    "python-multipart>=0.0.6,<0.1.0",
    "structlog>=23.2.0,<24.0.0",
    "keyring>=24.3.0,<25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0", 
    "httpx>=0.25.0,<0.26.0",
    "factory-boy>=3.3.0,<3.4.0",
    "coverage[toml]>=7.3.0,<8.0.0",
    "ruff>=0.1.6,<0.2.0",
    "black>=23.11.0,<24.0.0",
    "mypy>=1.7.0,<1.8.0",
    "types-python-jose>=3.3.4,<3.4.0",
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "B", "C4", "ARG", "SIM", "TCH", "PIE"]
ignore = ["E501", "B008", "ARG001"]

[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
no_implicit_optional = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests (large datasets, performance)",
]